(B) Fixed sized arrow disappearance

+++ branches/ff-heating-exp-1/ff_model_graphics.cpp
@@ -184,7 +184,7 @@ void ff_mgr_show_next_step()
     double tmag;
     double kvec;
     double space_k = 1E6;
-    long kSecondary = 1;
+    double kSecondary = 1;
     
 
     //step++;
